# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AIPT API",
        "description": "This is the Swagger UI documentation for AIPT API",
        "contact": {
            "email": "samirram007@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://aipt-api.local",
            "description": "Main API server"
        },
        {
            "url": "https://aipt-api.local",
            "description": "Main API server"
        }
    ],
    "paths": {
        "/api/account_groups": {
            "get": {
                "tags": [
                    "Account Group"
                ],
                "summary": "Get a list of account groups",
                "operationId": "3eb15e55da32543002969adfb9772e9c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AccountGroupResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Account Group"
                ],
                "summary": "Create a new account group",
                "operationId": "ac17ec873f38c1ca54d4a822ae6e866f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AccountGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/account_groups/{id}": {
            "get": {
                "tags": [
                    "Account Group"
                ],
                "summary": "Get a specific account group",
                "operationId": "7441517bd3fa90c533f72a5b682f1b69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the account group",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AccountGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountGroupRequest": {
                "properties": {
                    "name": {
                        "description": "The name of the account group",
                        "type": "string"
                    },
                    "code": {
                        "description": "The code of the account group",
                        "type": "string"
                    },
                    "account_nature_id": {
                        "description": "The ID of the account nature associated with the account group",
                        "type": "integer"
                    },
                    "description": {
                        "description": "The description of the account group",
                        "type": "string"
                    },
                    "status": {
                        "description": "The status of the account group",
                        "type": "boolean"
                    },
                    "icon": {
                        "description": "The icon of the account group",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AccountGroupResource": {
                "title": "AccountGroupResource",
                "description": "AccountGroupResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Account Group Name"
                    },
                    "code": {
                        "type": "string",
                        "example": "AG001"
                    },
                    "accountNatureId": {
                        "type": "integer",
                        "example": 1
                    },
                    "accountNature": {
                        "$ref": "#/components/schemas/AccountNatureResource"
                    },
                    "accountLedgers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountLedgerResource"
                        }
                    },
                    "description": {
                        "type": "string",
                        "example": "Account Group Description"
                    },
                    "status": {
                        "type": "string",
                        "example": "Active"
                    },
                    "icon": {
                        "type": "string",
                        "example": "fa fa-user"
                    }
                },
                "type": "object"
            },
            "AccountLedgerResource": {
                "title": "Account Ledger Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Example Account Ledger"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-23T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-23T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "AccountNatureResource": {
                "title": "Account Nature Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Example Account Nature"
                    },
                    "code": {
                        "type": "string",
                        "example": "EXAMPLE"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is an example account nature."
                    },
                    "account_groups": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountGroupResource"
                        }
                    },
                    "account_ledgers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountLedgerResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}